% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freud_line}
\alias{freud_line}
\title{Sigmund Freud complete work by lines}
\format{
# `freud_line`
A data frame with 125,571 rows and 6 columns:

\describe{
  \item{pg_all}{page number in original PDF file}
  \item{publi_yr}{year of book / article publication}
  \item{writ_yr}{year of book / article writing}
  \item{title}{title of book / article}
  \item{pg_title}{page number in book / article}
  \item{text}{raw text from each page}
}
}
\source{
<https://www.valas.fr/IMG/pdf/Freud_Complete_Works.pdf>
}
\usage{
freud_line
}
\description{
Sigmund Freud complete work by lines
}
\keyword{datasets}
