% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freud_page}
\alias{freud_page}
\title{Sigmund Freud data}
\format{
# `freud_page`
A data frame with 5,102 rows and 2 columns:
\describe{
  \item{pg_all}{page of original PDF file}
  \item{text}{raw text from each page}
  }
}
\source{
<https://www.valas.fr/IMG/pdf/Freud_Complete_Works.pdf>
}
\usage{
freud_page
}
\description{
Sigmund Freud complete work by pages
}
\keyword{datasets}
