% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freud_sntnce}
\alias{freud_sntnce}
\title{Tibble with Sigmund Freud's complete work by sentences}
\format{
`freud_sntnce` a data frame with 125,571 rows and 6 columns:

\describe{
  \item{pg_all}{integer, page number in original PDF file}
  \item{publi_yr}{integer, year of book / article publication}
  \item{writ_yr}{integer, year of book / article writing}
  \item{title}{factor, title of book / article}
  \item{pg_title}{integer, page number in book / article}
  \item{sentence}{character, raw text from each sentence}
}
}
\source{
\href{https://www.valas.fr/IMG/pdf/Freud_Complete_Works.pdf}{Text source}
}
\usage{
freud_sntnce
}
\description{
Tibble with Sigmund Freud's complete work by sentences
}
\keyword{datasets}
