% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freud_word}
\alias{freud_word}
\title{#' Sigmund Freud complete work by words}
\format{
## `freud_word`
A data frame with 125,571 rows and 6 columns:

\describe{
  \item{pg_all}{page number in original PDF file}
  \item{publi_yr}{year of book / article publication}
  \item{writ_yr}{year of book / article writing}
  \item{title}{title of book / article}
  \item{pg_title}{page number in book / article}
  \item{wor}{word from each page, without stop words}
  ...
}
}
\source{
<https://www.valas.fr/IMG/pdf/Freud_Complete_Works.pdf>
}
\usage{
freud_word
}
\description{
Sigmund Freud complete work by word, stop words have been removed using
tidytext::stop_words according to lexicon onix, SMART, and snowball
}
\keyword{datasets}
