% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freud_page}
\alias{freud_page}
\title{Tibble with Sigmund Freud's complete work by page}
\format{
`freud_page` a data frame with 5,102 rows and 2 columns:
\describe{
  \item{pg_all}{integer, page of original PDF file}
  \item{text}{character, raw text from each page}
  }
}
\source{
\href{https://www.valas.fr/IMG/pdf/Freud_Complete_Works.pdf}{Text source}
}
\usage{
freud_page
}
\description{
Sigmund Freud complete work by pages
}
\keyword{datasets}
